Conflicts:
	platform/lang-impl/src/com/intellij/codeInsight/actions/AbstractLayoutCodeProcessor.java

+++ platform/util/src/com/intellij/util/Restarter.java
@@ -155,6 +155,7 @@ public class Restarter {
 
   public static File createTempExecutable(File executable) throws IOException {
     File executableDir = new File(System.getProperty("user.home") + "/." + System.getProperty("idea.paths.selector") + "/restart");
+    if (!FileUtilRt.createDirectory(executableDir)) throw new IOException("Cannot create dir: " + executableDir);
     File copy = new File(executableDir.getPath() + "/" + executable.getName());
     if (!FileUtilRt.ensureCanCreateFile(copy) || (copy.exists() && !copy.delete())) {
        String ext = FileUtilRt.getExtension(executable.getName());
